==> notebooks/3.2 PD alpha effect size (r - RDA).ipynb <==
library(vegan)

mf = read.csv('../data/mapping_PDalpha.txt', stringsAsFactors=FALSE, sep='\t')
colnames(mf)[1] = 'SampleName'

head(mf)

dim(mf)

# convert categorial to factors
vars_cat = c('Sex','GIERACE', 'SITE', 'TUDRAMT', 'TURSMOKE', 'M1ADEPR', 'M1VITMND', 'M1ANTIB', 'M1PROBI', 
               'OHSEAS', 'VDstatus', 'OHV1D2CT', 'OHVD2CT')
mf[vars_cat] = lapply(mf[vars_cat], factor)

# convert continuous to numeric
vars_cts = c('Latitude', 'Longitude', 'Age', 'Height', 'Weight', 'BMI', 'PASCORE', 'DTVITD', 
             'OHV1D3', 'OHV24D3', 'OHVD3', 'OHVD2', 'OHV1D2', 'OHVDTOT', 'OHV1DTOT', 'alpha_pd')

length(vars_cat)
length(vars_cts)

dat = mf[c(vars_cat, vars_cts)]
dim(dat)
head(dat)

summary(dat)

dat = dat[complete.cases(dat), ]
print(dim(dat))
alpha = dat$alpha_pd
dat = dat[, -which(names(dat) %in% 'alpha_pd')]
dim(dat)

mod0 <- rda(alpha ~ 1., dat)  # Model with intercept only
mod1 <- rda(alpha ~ ., dat)  # Model with all explanatory variables

step.res <- ordiR2step(mod0, mod1, perm.max = 1000)

table = step.res$anova
table

# Step: R2.adj= 0 
# Call: alpha ~ 1 
 
#                   R2.adjusted
# <All variables>  0.1259630505
# + OHV1DTOT       0.0455455639
# + SITE           0.0311618389
# + GIERACE        0.0271660306
# + M1ANTIB        0.0211519590
# + OHVDTOT       -0.0012109491

# explain negative adjusted R^2: https://stats.stackexchange.com/questions/183265/what-does-negative-r-squared-mean

table$ES.independent = c(0.0455455639, 0.0311618389, 0.0271660306, 0.0012109491, 0.0211519590, 0.1259630505)
table$ES.RDA = c(table$R2.adj[1], table$R2.adj[2]-table$R2.adj[1], 
                       table$R2.adj[3]-table$R2.adj[2], table$R2.adj[4]-table$R2.adj[3],
                       table$R2.adj[5]-table$R2.adj[4], table$R2.adj[6]-table$R2.adj[5])
table = table[-6, ]

step.res$call

table
